filter 기능 + DB 버전 추가

+++ backend/src/modules/user/user.controller.ts
@@ -9,10 +9,13 @@ import {
   Delete,
   Body,
   BadRequestException,
+  UnauthorizedException,
+  Req,
 } from '@nestjs/common';
 import { LoggedInGuard } from '@/guards/logged-in.guard';
 import { UserService } from './user.service';
 import { UserCreateDto, UserUpdateDto } from '@/dto/user.dto';
+import { customRequest } from '@/interface/user.interface';
 
 @Controller('user')
 export class UserController {
@@ -34,9 +37,9 @@ export class UserController {
 
   @Get(':user_id')
   async getOneUser(@Param('user_id') user_id: string) {
+    console.log('[user.controller:getOneUser] starting function');
+    console.log('[user.controller:getOneUser] user_id: ', user_id);
     try {
-      console.log('[user.controller:getOneUser] starting function');
-      console.log('[user.controller:getOneUser] user_id: ', user_id);
       const res = await this.userService.getUserByKey(user_id);
       console.log('[user.controller:getOneUser] res: ', res);
       return res;
@@ -48,9 +51,9 @@ export class UserController {
 
   @Post()
   async createUser(@Body() data: UserCreateDto) {
+    console.log('[user.controller:createUser] starting function');
+    console.log('[user.controller:createUser] data: ', data);
     try {
-      console.log('[user.controller:createUser] starting function');
-      console.log('[user.controller:createUser] data: ', data);
       const res = await this.userService.createUser(data);
       console.log('[user.controller:createUser] res: ', res);
       return res;
@@ -65,11 +68,18 @@ export class UserController {
   async putOneUser(
     @Param('user_id') user_id: string,
     @Body() putUserBody: UserUpdateDto,
+    @Req() req: customRequest,
   ) {
+    console.log('[user.controller:putOneUser] starting function');
+    console.log('[user.controller:putOneUser] user_id: ', user_id);
+    console.log('[user.controller:putOneUser] putUserBody: ', putUserBody);
+    if (req.user.user_id !== user_id) {
+      console.log(
+        '[user.controller:putOneUser] user_id is not same as req.user.user_id',
+      );
+      throw new UnauthorizedException();
+    }
     try {
-      console.log('[user.controller:putOneUser] starting function');
-      console.log('[user.controller:putOneUser] user_id: ', user_id);
-      console.log('[user.controller:putOneUser] putUserBody: ', putUserBody);
       const res = await this.userService.putOneUser(user_id, putUserBody);
       console.log('[user.controller:putOneUser] res: ', res);
       return res;
@@ -81,10 +91,19 @@ export class UserController {
 
   @Delete(':user_id')
   @UseGuards(LoggedInGuard)
-  async deleteOneUser(@Param('user_id') user_id: string) {
+  async deleteOneUser(
+    @Param('user_id') user_id: string,
+    @Req() req: customRequest,
+  ) {
+    console.log('[user.controller:deleteOneUser] starting function');
+    console.log('[user.controller:deleteOneUser] user_id: ', user_id);
+    if (req.user.user_id !== user_id) {
+      console.log(
+        '[user.controller:deleteOneUser] user_id is not same as req.user.user_id',
+      );
+      throw new UnauthorizedException();
+    }
     try {
-      console.log('[user.controller:deleteOneUser] starting function');
-      console.log('[user.controller:deleteOneUser] user_id: ', user_id);
       const res = await this.userService.deleteOneUser(user_id);
       console.log('[user.controller:deleteOneUser] res: ', res);
       return res;
